requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
base58==1.0.3
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
PyNaCl==1.5.0
pysha3==1.0.2
python-dotenv==1.0.1
python-rapidjson-schema==0.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
